# swaggerDoc.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple CRUD server",
        "version": "1.0.0",
        "title": "Swagger Post & Comment",
        "contact": {
            "name": "Prathamesh Patil",
            "email": "prathmeshp658@gmail.com",
            "url": "www.iampratham.com"
        }
    },
...
    "basePath": "/api/",
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "jwt": []
        }
    ],
    "paths": {
...
        "/posts/": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summery": "List of posts",
                "description": "Returns A list of Posts created by users ",
                "operationId": "getAllPost",
                "produces": [
                    "appliation/json"
                ],
                "parameters": [],
                "responses": {
                    "401": {
                        "description": "Unauthorized or invalid token",
                        "schema": {
                            "$ref": "#/components/responses/common"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/components/responses/common"
                        }
                    },
                    "404": {
                        "description": "Returns a post doesn't exists error with given credentials.",
                        "schema": {
                            "$ref": "#/components/responses/common"
                        }
                    },
                    "200": {
                        "description": "Returns the array of posts.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "user": {
                                                "type": "object",
                                                "properties": {
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "username": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "comments": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object",
                                                    "properties": {
                                                        "user": {
                                                            "type": "object",
                                                            "properties": {
                                                                "username": {
                                                                    "type": "string"
                                                                }
                                                            }
                                                        },
                                                        "_id": {
                                                            "type": "string"
                                                        },
                                                        "comment": {
                                                            "type": "string"
                                                        }
                                                    }
                                                }
                                            },
                                            "_id": {
                                                "type": "string"
                                            },
                                            "title": {
                                                "type": "string"
                                            },
                                            "content": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summery": "Inserts a new post",
                "description": "Creates a new post in database",
                "operationId": "createPost",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created Post object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "title": {
                                    "type": "string"
                                },
                                "content": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created Post inside database.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "object",
                                    "properties": {
                                        "user": {
                                            "type": "string"
                                        },
                                        "_id": {
                                            "type": "string"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "content": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorised or invalid token.",
                        "schema": {
                            "$ref": "#/components/responses/common"
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "$ref": "#/components/responses/common"
                        }
                    }
                }
            }
        },
...
        "/comments/": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summery": "Get a List of comments",
                "description": "Returns a list of comments made on posts by users",
                "operationId": "getComments",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "401": {
                        "description": "Unauthorised or invalid token.",
                        "schema": {
                            "$ref": "#/components/responses/common"
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "schema": {
                            "$ref": "#/components/responses/common"
                        }
                    },
                    "404": {
                        "description": "Invalid post error.",
                        "schema": {
                            "$ref": "#/components/responses/common"
                        }
                    },
                    "400": {
                        "description": "Validation error.",
                        "schema": {
                            "$ref": "#/components/responses/common"
                        }
                    },
                    "200": {
                        "description": "Returns a list of comments with populated data.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "user": {
                                                "type": "object",
                                                "properties": {
                                                    "_id": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "username": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "post": {
                                                "_id": {
                                                    "type": "string"
                                                },
                                                "title": {
                                                    "type": "string"
                                                },
                                                "content": {
                                                    "type": "string"
                                                }
                                            },
                                            "_id": {
                                                "type": "string"
                                            },
                                            "comment": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Comment"
                ],
                "summery": "Create a new Comment ",
                "description": "Create a new Comment in db",
                "operationId": "createComment",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "description": "An Comment object to be Inserted",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "post": {
                                    "type": "string"
                                },
                                "comment": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a comment object just made by user.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "object",
                                    "properties": {
                                        "user": {
                                            "type": "string"
                                        },
                                        "post": {
                                            "type": "string"
                                        },
                                        "_id": {
                                            "type": "string"
                                        },
                                        "comment": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "components": {
        "responses": {
            "common": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            }
        }
    }
}